feat: Add SignedNumeric protocol.
