+++ Sources/SternBroctTreeSwift/Core/Fraction.swift
@@ -107,119 +107,6 @@ public protocol Fraction : SBTreeNode, SignedNumeric, Comparable, Hashable, Cust
 
 }
 
-/// Default implementation for SBTreeNode.
-extension Fraction where Number : SignedInteger {
-
-    /// Returns a boolean value whether this and the other are adjacent.
-    ///
-    /// - Parameter other: The other concrete rational to determine adjacent.
-    /// - Returns: The two values are adjacent or not.
-    public func isAdjacent(to other: Self) -> Bool {
-        let ad = Int64(numerator * other.denominator)
-        let bc = Int64(denominator * other.numerator)
-        return abs(ad - bc) == 1
-    }
-
-    /// Returns simplicity of a rational.
-    ///
-    /// if **r=a/b** is in reduced form, **the simplicity of r** is defined to be **L(r)≡1/ab**.
-    public var simplicity: Self {
-        let ab = Int64(numerator * denominator)
-        return Self("1/\(ab)")
-    }
-
-    /// Returns total of a rational.
-    ///
-    /// if **r=a/b** is in reduced form, define the total of r to be **t(r) ≡ a+b**.
-    public var total: Number {
-        return numerator + denominator
-    }
-
-    /// Returns the interger value of mixed rational.
-    public var mixedPart: Number {
-        return numerator / denominator
-    }
-
-    /// Returns the numerator of mixed rational.
-    public var mixedRemainder: Number {
-        return numerator % denominator
-    }
-
-}
-
-extension Fraction {
-
-    /// Returns an new insntance from random numer and denom.
-    ///
-    /// The denominator must not be zero.
-    ///
-    /// - Parameter max: The maximum value of numer and denom.
-    /// - Returns: A concrete rational type is instanciated by random values.
-    public static func random(max: Int32 = 1000000) -> Self {
-        let numerator = Int32.random(in: 0...max)
-        let denominator = Int32.random(in: 1...max)
-        return Self("\(numerator)/\(denominator)")
-    }
-    
-}
-
-/// Default implementation for Equtable.
-extension Fraction where Number : SignedInteger {
-
-    public static func == (lhs: Self, rhs: Self) -> Bool {
-        return lhs.numerator == rhs.numerator && lhs.denominator == rhs.denominator
-    }
-
-}
-
-/// Default implementation for Comparable.
-extension Fraction where Number : SignedInteger {
-
-    public static func < (lhs: Self, rhs: Self) -> Bool {
-
-        // If a, b, c, and d are positive, the result of a/b < c/d can represent as ad < bc.
-        // 1. To remove left side devider, let both sides multiplied by b. (a < bc/d)
-        // 2. To remove right side devider, let both sides multipied by d. (ad < bc)
-
-        let a = lhs.numerator
-        let b = lhs.denominator
-        let c = rhs.numerator
-        let d = rhs.denominator
-
-        let ad: Int64 = Int64(a * d)
-        let bc: Int64 = Int64(b * c)
-
-        return ad < bc
-    }
-
-}
-
-/// Default implementation for Hashable.
-extension Fraction where Number : SignedInteger {
-
-    public func hash(into hasher: inout Hasher) {
-
-        // In reduced form, SBTree node's fruction must be identified in the tree.
-        let x = simplified()
-        hasher.combine(x.numerator)
-        hasher.combine(x.denominator)
-    }
-
-}
-
-/// Default implementation for Convertibles.
-extension Fraction where Number : SignedInteger {
-
-    public var description: String { "\(numerator)/\(denominator)" }
-
-    public var float32Value: Float32 { Float32(numerator) / Float32(denominator) }
-
-    public var float64Value: Float64 { Float64(numerator) / Float64(denominator) }
-
-    public var doubleValue: Double { Double(numerator) / Double(denominator) }
-
-}
-
 extension Fraction where Number == Int {
 
     public var decimalValue: Decimal { Decimal(numerator) / Decimal(denominator) }
@@ -249,62 +136,3 @@ extension Fraction where Number == Int64 {
     public var decimalValue: Decimal { Decimal(numerator) / Decimal(denominator) }
 
 }
-
-/// Default implementation for SignedNumeric.
-extension Fraction where Number : SignedInteger, Magnitude == Double, IntegerLiteralType == Number {
-
-    public static var zero: Self {
-        return Self("0/1")
-    }
-
-    public var magnitude: Double {
-        return abs(Double(numerator)/Double(denominator))
-    }
-
-    public init?<T>(exactly source: T) where T : BinaryInteger {
-        self.init(numerator: Number(source), denominator: 1)
-    }
-
-    public init(integerLiteral value: Number) {
-        self.init(numerator: value, denominator: 1)
-    }
-
-    public static func + (lhs: Self, rhs: Self) -> Self {
-        let denominator = lhs.denominator * rhs.denominator
-        let numerator = (lhs.numerator * rhs.denominator) + (rhs.numerator * lhs.denominator)
-        return self.init(numerator: numerator, denominator: denominator)
-    }
-
-    public static func - (lhs: Self, rhs: Self) -> Self {
-        let denominator = lhs.denominator * rhs.denominator
-        let numerator = (lhs.numerator * rhs.denominator) - (rhs.numerator * lhs.denominator)
-        return self.init(numerator: numerator, denominator: denominator)
-    }
-
-    public static func * (lhs: Self, rhs: Self) -> Self {
-        return self.init(numerator: lhs.numerator * rhs.numerator, denominator: lhs.denominator * rhs.denominator)
-    }
-
-    public static func *= (lhs: inout Self, rhs: Self) {
-        lhs = lhs * rhs
-    }
-
-    /// Returns the quotient of dividing the first value by the second.
-    ///
-    /// - Parameters:
-    ///   - lhs: The value to divide.
-    ///   - rhs: The value to divide lhs by. rhs must not be zeo.
-    public static func / (lhs: Self, rhs: Self) -> Self {
-        return lhs * self.init(numerator: rhs.denominator, denominator: rhs.numerator)
-    }
-
-    /// Divides the first value by the second and stores the quotient in the left-hand-side variable
-    /// - Parameters:
-    ///   - lhs: The value to divide.
-    ///   - rhs: The value to divide lhs by. rhs must not be zeo.
-    public static func /= (lhs: inout Self, rhs: Self) {
-        lhs = lhs / rhs
-    }
-
-}
-
